`bool load_SRTM_world`: `GLfloat` -> `double`.
Sijainti pituus- ja leveyspiireinä on hyvä ilmoittaa tarkasti. Sitä
paitsi OpenGL:lle ei muutenkaan välitetä pallokoordinaatiston
koordinaatteja, vaan OpenGL:lle välitetään ainoastaan suorakulmaisen
koordinaatiston koordinaatteja.

	modified:   common/heightmap_loader.cpp
	modified:   common/heightmap_loader.hpp

+++ common/heightmap_loader.hpp
@@ -12,8 +12,8 @@ namespace model
 
     bool load_SRTM_world(
             std::string image_path,
-            GLfloat southern_latitude,
-            GLfloat western_longitude,
+            double southern_latitude,
+            double western_longitude,
             std::vector<glm::vec3> &out_vertices,
             std::vector<glm::vec2> &out_UVs,
             std::vector<glm::vec3> &out_normals);
